feat(issue-9): Implemented project restructure and refactor functions.
