+++ tests/test_app.py
@@ -1,63 +1,4 @@
-from fast_zero.schemas import UserPublic
-
-
 def test_root_deve_retornar_200_e_ola_mundo(client):
     response = client.get('/')  # Act
     assert response.status_code == 200  # Assert
     assert response.json() == {'message': 'Olá Mundo!'}  # Assert
-
-
-def test_create_user(client):
-    response = client.post(
-        '/users/',
-        json={
-            'username': 'alice',
-            'email': 'alice@example.com',
-            'password': 'secret',
-        },
-    )
-    assert response.status_code == 201
-    assert response.json() == {
-        'username': 'alice',
-        'email': 'alice@example.com',
-        'id': 1,
-    }
-
-
-def test_read_users(client):
-    response = client.get('/users')
-    assert response.status_code == 200
-    assert response.json() == {'users': []}
-
-
-def test_read_users_with_users(client, user):
-    user_schema = UserPublic.model_validate(user).model_dump()
-    response = client.get('/users/')
-    assert response.json() == {'users': [user_schema]}
-
-
-def test_update_user(client, user, token):
-    response = client.put(
-        f'/users/{user.id}',
-        headers={'Authorization': f'Bearer {token}'},
-        json={
-            'username': 'bob',
-            'email': 'bob@example.com',
-            'password': 'mynewpassword',
-        },
-    )
-    assert response.status_code == 200
-    assert response.json() == {
-        'username': 'bob',
-        'email': 'bob@example.com',
-        'id': 1,
-    }
-
-
-def test_delete_user(client, user, token):
-    response = client.delete(
-        f'/users/{user.id}',
-        headers={'Authorization': f'Bearer {token}'},
-    )
-    assert response.status_code == 200
-    assert response.json() == {'detail': 'User deleted'}
