--- tests/test_security.py
@@ -1,25 +1,0 @@
-from jose import jwt
-
-from fast_zero.security import create_access_token, SECRET_KEY
-
-
-def test_jwt():
-    data = {'test': 'test'}
-    token = create_access_token(data)
-
-    decoded = jwt.decode(token, SECRET_KEY, algorithms=['HS256'])
-
-    assert decoded['test'] == data['test']
-    assert decoded['exp']  # Testa se o valor de exp foi adicionado ao token
-
-
-def test_get_token(client, user):
-    response = client.post(
-        '/token',
-        data={'username': user.email, 'password': user.clean_password},
-    )
-    token = response.json()
-
-    assert response.status_code == 200
-    assert 'access_token' in token
-    assert 'token_type' in token
